[refactor] 删除部分Converter
