+++ FluentWeather.Uwp/App.xaml
@@ -13,8 +13,6 @@
                 <themes:Generic/>
                 <themes:Theme />
             </ResourceDictionary.MergedDictionaries>
-            <valueConverters:WeatherTypeToIconConverter x:Key="WeatherTypeToIconConverter" />
-            <valueConverters:WeatherToIconConverter x:Key="WeatherToIconConverter" />
             <valueConverters2:BoolNegationConverter x:Key="BoolNegationConverter" />
             <valueConverters2:ValueConverterGroup x:Name="NullToVisibilityConverter">
                 <valueConverters2:NullToBoolConverter IsInverted="True" />
@@ -26,23 +24,10 @@
             </valueConverters2:ValueConverterGroup>
             <valueConverters:VisibilityInverter x:Name="VisibilityInverter" />
             <valueConverters2:EnumToBoolConverter x:Name="EnumToBoolConverter" />
-            <valueConverters:WindDirectionToDescriptionConverter x:Name="WindDirectionToDescriptionConverter" />
             <valueConverters2:NullToBoolConverter x:Name="NullToBoolConverter" IsInverted="True" />
             <valueConverters:AutoSuggestSuggestionParameterConverter x:Key="AutoSuggestSuggestionParameterConverter" />
-            <valueConverters:PrecipitationProbabilityToVisibilityConverter x:Key="PrecipitationProbabilityToVisibilityConverter" />
-            <valueConverters:WindScaleToDescriptionConverter x:Key="WindScaleToDescriptionConverter" />
             <valueConverters:DateConverter x:Key="DateConverter" />
             <valueConverters:ShortDateConverter x:Key="ShortDateConverter" />
-            <Style x:Name="RepositionControlStyle" TargetType="FrameworkElement">
-                <Setter Property="Transitions">
-                    <Setter.Value>
-                        <TransitionCollection>
-                            <RepositionThemeTransition IsStaggeringEnabled="False" />
-                        </TransitionCollection>
-                    </Setter.Value>
-                </Setter>
-            </Style>
-
         </ResourceDictionary>
     </Application.Resources>
 </Application>
