+++ FluentWeather.Uwp/Helpers/ValueConverters/ConverterMethods.cs
@@ -7,6 +7,7 @@ using Windows.UI.Xaml.Media;
 using FluentWeather.Abstraction.Models;
 using Windows.Devices.Geolocation;
 using Newtonsoft.Json.Linq;
+using Windows.ApplicationModel.Resources;
 
 namespace FluentWeather.Uwp.Helpers.ValueConverters;
 
@@ -17,6 +18,18 @@ public static class ConverterMethods
         var series = presenter as CategoricalStrokedSeries;
         return series?.Stroke;
     }
-
+    public static string GetWindDirectionDescription(WindDirection dir)
+    {
+        return ResourceLoader.GetForCurrentView().GetString("WindDirection_" + dir);
+    }
+    public static string GetWindScaleDescription(string scale)
+    {
+        if (scale.Contains("-"))
+        {
+            var s = scale.Split("-");
+            scale = s[1];
+        }
+        return ResourceLoader.GetForCurrentView().GetString("WindScaleDescription_" + scale);
+    }
 
 }
