--- FluentWeather.Uwp/Helpers/ValueConverters/PrecipitationProbabilityToVisibilityConverter.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Data;
-
-namespace FluentWeather.Uwp.Helpers.ValueConverters;
-
-public sealed class PrecipitationProbabilityToVisibilityConverter : IValueConverter
-{
-    public object Convert(object value, Type targetType, object parameter, string language)
-    {
-        if (value is not int precip) return null;
-        return (precip >= 10) ? Visibility.Visible : Visibility.Collapsed;
-    }
-
-    public object ConvertBack(object value, Type targetType, object parameter, string language)
-    {
-        throw new NotImplementedException();
-    }
-}
