+++ FluentWeather.Uwp/Pages/MainPage.xaml
@@ -9,6 +9,7 @@
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
     xmlns:controls1="using:FluentWeather.Uwp.Controls"
     xmlns:converters="using:ValueConverters"
+    xmlns:converters1="using:FluentWeather.Uwp.Helpers.ValueConverters"
     xmlns:core="using:Microsoft.Xaml.Interactions.Core"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:dialogs="using:FluentWeather.Uwp.Controls.Dialogs"
@@ -217,7 +218,7 @@
                                     FontSize="18"
                                     Foreground="{ThemeResource SystemControlForegroundBaseMediumBrush}">
                                     <Run x:Uid="FeelsLike" />
-                                    <Run Text="{Binding WeatherNow.ApparentTemperature}" /><Run Text="°" />
+                                    <Run Text="{x:Bind ViewModel.WeatherNow.ApparentTemperature}" /><Run Text="°" />
                                 </TextBlock>
                                 <Button
                                     Margin="12,0,0,0"
@@ -561,7 +562,7 @@
                                     <TextBlock
                                         FontSize="18"
                                         Foreground="{ThemeResource SystemControlForegroundBaseMediumBrush}"
-                                        Text="{x:Bind ViewModel.WeatherNow.WindDirection, Converter={StaticResource WindDirectionToDescriptionConverter}}" />
+                                        Text="{x:Bind converters1:ConverterMethods.GetWindDirectionDescription(ViewModel.WeatherNow.WindDirection)}" />
                                     <TextBlock FontSize="25">
                                         <Run FontSize="25" Text="{x:Bind ViewModel.WeatherNow.WindScale}" />
                                         <Run x:Uid="WindScaleText" FontSize="16" />
@@ -570,7 +571,7 @@
                                         <Flyout>
                                             <StackPanel MaxWidth="256">
                                                 <TextBlock FontSize="18" Foreground="{ThemeResource SystemControlForegroundBaseMediumBrush}">
-                                                    <Run Text="{x:Bind ViewModel.WeatherNow.WindDirection, Converter={StaticResource WindDirectionToDescriptionConverter}}" />
+                                                    <Run Text="{x:Bind converters1:ConverterMethods.GetWindDirectionDescription(ViewModel.WeatherNow.WindDirection)}" />
                                                     <Run Text="{x:Bind ViewModel.WeatherNow.WindScale}" />
                                                     <Run x:Uid="WindScaleText" FontSize="14" />
                                                 </TextBlock>
@@ -581,7 +582,7 @@
                                                 <TextBlock
                                                     FontSize="16"
                                                     IsTextSelectionEnabled="True"
-                                                    Text="{x:Bind ViewModel.WeatherNow.WindScale, Converter={StaticResource WindScaleToDescriptionConverter}}"
+                                                    Text="{x:Bind converters1:ConverterMethods.GetWindScaleDescription(ViewModel.WeatherNow.WindScale)}"
                                                     TextWrapping="Wrap" />
                                             </StackPanel>
                                         </Flyout>
